Implementation of Spring Boot Application: - Added com.example directory to hold all Spring files - Added MiniSurveyMonkeyApplication.java (@SpringBootApplication) - Moved all classes into com.example package
