+++ src/main/java/com/example/MiniSurveyMonkeyApplication.java
@@ -1,0 +1,11 @@
+package com.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MiniSurveyMonkeyApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(MiniSurveyMonkeyApplication.class, args);
+    }
+}
