+++ src/main/java/com/example/SurveryRepository.java
@@ -1,3 +1,5 @@
+package com.example;
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
