+++ src/main/java/com/example/Survey.java
@@ -1,17 +1,19 @@
+package com.example;
+
 import jakarta.persistence.*;
 
 @Entity
-public class SurveyResponse {
+public class Survey {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false)
     private Long id;
 
-    public SurveyResponse(Long id) {
+    public Survey(Long id) {
         this.id = id;
     }
 
-    public SurveyResponse() {
+    public Survey() {
 
     }
