+++ src/main/java/com/example/SurveyController.java
@@ -1,3 +1,5 @@
+package com.example;
+
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
