+++ src/main/java/com/example/SurveyResult.java
@@ -1,17 +1,20 @@
+package com.example;
+
 import jakarta.persistence.*;
 
 @Entity
-public class SurveyQuestion {
+@Table(name = "survey_results")
+public class SurveyResult {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false)
     private Long id;
 
-    public SurveyQuestion(Long id) {
+    public SurveyResult(Long id) {
         this.id = id;
     }
 
-    public SurveyQuestion() {
+    public SurveyResult() {
 
     }
